Refactor App presentation component into Presenter
