+++ app/frontend/src/Presenter/index.css
@@ -1,0 +1,22 @@
+.presenter .controller-container {
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  height: 273px;
+  width: 416px;
+}
+
+.presenter .controller-container .expand-icon {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  height: 48px;
+}
+
+.presenter .controller-container.fullscreen {
+  top: 0;
+  left: 0;
+  height: 100%;
+  width: 100%;
+  font-size: 3vmin;
+}
